ml_analisis_imagen/requirements.txt
# Framework web
fastapi
uvicorn[standard]
python-multipart

# Machine Learning
scikit-learn
numpy
pandas
joblib

# Computer Vision
opencv-python
Pillow

# Utilidades
python-dotenv
pydantic
pydantic-settings
